% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeFunGF.R
\name{calcGF}
\alias{calcGF}
\title{Calcualte G-formula-based Causal Effect Estimates}
\usage{
calcGF(data, outcome_model_obj, treatment_var_name)
}
\arguments{
\item{data}{the dataframe. Will be coerced from "tbl_df" to data.frame.}

\item{outcome_model_obj}{The fitted model object (usually a glm).}

\item{treatment_var_name}{the name of the treatment variable. Must line up with formula.}
}
\description{
Calcualte G-formula-based Causal Effect Estimates
}
