% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDR.R
\name{eeFunDRIPTW}
\alias{eeFunDRIPTW}
\title{Estimating Function for DRIPTW}
\usage{
eeFunDRIPTW(data, outcome_model_obj, trt_model_obj, num_outcome_params,
  num_treatment_params, idx_outcome_params, idx_treatment_params,
  outcome_var_name, treatment_var_name, calcFunDRIPTW, predictOutcome,
  predictTreatment)
}
\arguments{
\item{data}{the dataframe. Will be coerced from "tbl_df" to data.frame.}

\item{outcome_model_obj}{The fitted outcome model object (usually a glm).}

\item{trt_model_obj}{The fitted treatment model object (usually a glm).}

\item{num_outcome_params}{n_trt}

\item{num_treatment_params}{n_trt}

\item{idx_outcome_params}{vector of indices}

\item{idx_treatment_params}{vector of indices}

\item{outcome_var_name}{The name of the column in the dataframe indicating outcome of interest}

\item{treatment_var_name}{The name of the column in the dataframe indicating treatment}

\item{calcFunDRIPTW}{this is a function object specified in the weight_type argument}

\item{predictOutcome}{the function to predict outcome probs}

\item{predictTreatment}{the function to predict propensity score}
}
\description{
This function is to be passed into geex::m_estimate
}
