% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateGF.R
\name{eeFunGF}
\alias{eeFunGF}
\title{Estimating Function for IPTW}
\usage{
eeFunGF(data, outcome_model_obj, predictOutcome, treatment_var_name)
}
\arguments{
\item{data}{the dataframe. Will be coerced from "tbl_df" to data.frame.}

\item{outcome_model_obj}{The fitted model object (usually a glm).}

\item{predictOutcome}{the function to predict outcome probs}

\item{treatment_var_name}{the name of the treatment variable. Must line up with formula.}
}
\description{
This function is to be passed into geex::m_estimate
}
