% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeFunIPW.R
\name{eeFunIPTW}
\alias{eeFunIPTW}
\title{Estimating Function for IPTW}
\usage{
eeFunIPTW(data, trt_model_obj, outcome_var_name, calcFunIPTW)
}
\arguments{
\item{data}{the dataframe. Will be coerced from "tbl_df" to data.frame.}

\item{trt_model_obj}{The fitted model object (usually a glm).}

\item{outcome_var_name}{The name of the column in the dataframe indicating outcome of interest}

\item{calcFunIPTW}{this is a function object specified in the weight_type argument}
}
\description{
This function is to be passed into geex::m_estimate
}
