% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDR.R
\name{estimateDRIPTW}
\alias{estimateDRIPTW}
\title{Estimate ATE with IPTW}
\usage{
estimateDRIPTW(data, outcome_formula, outcome_model_method = "logistic",
  treatment_formula, treatment_model_method = "logistic",
  weight_type = "unstabilized", ...)
}
\arguments{
\item{data}{the dataframe. Will be coerced from "tbl_df" to data.frame.}

\item{outcome_formula}{Two part formula: Outcome ~ Covars + Trt}

\item{outcome_model_method}{currently only supported "logistic" for logit-link binomial GLM.}

\item{treatment_formula}{Two part formula: Trt ~ Covars}

\item{treatment_model_method}{currently only supported "logistic" for logit-link binomial GLM.}

\item{weight_type}{Currently only supports "unstabilized"}

\item{...}{additional args}
}
\description{
Fits a parametric model and estimates ATE via IPTW with Wald-type confidence
intervals from the empirical sandwich standard error estimates.
}
