% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateGF.R
\name{estimateGF}
\alias{estimateGF}
\title{Estimate ATE with G-Formula}
\usage{
estimateGF(data, formula, model_method = "logistic", treatment_var_name, ...)
}
\arguments{
\item{data}{The dataframe. This will be coerced from "tbl_df" to data.frame.}

\item{formula}{Two-part formula: Outcome ~ Treatment + Covariates.}

\item{model_method}{Currently only supported "logistic" for logit-link binomial GLM.}

\item{treatment_var_name}{The name of the treatment variable's column in `data`}

\item{...}{additional arguments, especially those passed to
\code{\link[geex]{m_estimate}} or the modeling function}
}
\description{
Fits a parametric model and estimates ATE via IPTW with Wald-type confidence
intervals from the empirical sandwich standard error estimates.
}
