% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateIPTW.R
\name{estimateIPTW}
\alias{estimateIPTW}
\title{Estimate ATE with IPTW}
\usage{
estimateIPTW(data, formula, model_method = "logistic",
  weight_type = "unstabilized", ...)
}
\arguments{
\item{data}{The dataframe. This will be coerced from "tbl_df" to data.frame.}

\item{formula}{Three-part formula: Outcome | Treatment ~ Covars.}

\item{model_method}{Currently only supports "logistic" for logit-link binomial GLM.}

\item{weight_type}{Currently only supports "unstabilized"}

\item{...}{additional arguments, especially those passed to
\code{\link[geex]{m_estimate}} or the modeling function}
}
\description{
Fits a parametric model and estimates ATE via IPTW with Wald-type confidence
intervals from the empirical sandwich standard error estimates.
}
