% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{fitTreatmentModel}
\alias{fitTreatmentModel}
\title{Prep Outcome Model Matrices for Prediction}
\usage{
fitTreatmentModel(data, formula, model_method)
}
\arguments{
\item{data}{the dataframe. Will be coerced from "tbl_df" to data.frame.}

\item{formula}{Three-part formula: Outcome | Treatment ~ model_predictors. Will be coerced to object of type Formula.}

\item{model_method}{currently only supported "logistic" for logit-link binomial GLM.}
}
\description{
Prep Outcome Model Matrices for Prediction
}
