% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabgeexfuns.R
\name{grab_psiFUN_glm}
\alias{grab_psiFUN_glm}
\title{Temporary Function Until geex>1.0.3}
\usage{
grab_psiFUN_glm(object, data, weights = 1, ...)
}
\arguments{
\item{object}{stats::glm model object}

\item{data}{data}

\item{weights}{1}

\item{...}{more stuffs}
}
\description{
Temporary Function Until geex>1.0.3
}
