% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{makeOutcomeModelMats}
\alias{makeOutcomeModelMats}
\title{Prep Outcome Model Matrices for Prediction}
\usage{
makeOutcomeModelMats(data, outcome_model_obj, treatment_var_name)
}
\arguments{
\item{data}{the dataframe. Will be coerced from "tbl_df" to data.frame.}

\item{outcome_model_obj}{The fitted outcome model object (usually a glm).}

\item{treatment_var_name}{The name of the column in the dataframe indicating treatment}
}
\description{
Prep Outcome Model Matrices for Prediction
}
