% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDR.R
\name{unstabilizedDRIPTW}
\alias{unstabilizedDRIPTW}
\title{Function to Compute Horwitz-Thompson}
\usage{
unstabilizedDRIPTW(outcome, treatment, prob_treated, prob_outcome1,
  prob_outcome0, ...)
}
\arguments{
\item{outcome}{vector of outcome values}

\item{treatment}{vector of treatment values}

\item{prob_treated}{vector of propensity scores}

\item{prob_outcome1}{vector of E(Y=1 | Trt=1, covars)}

\item{prob_outcome0}{vector of E(Y=1 | Trt=0, covars)}

\item{...}{dots}
}
\description{
Computes estimates for each individual by IPTW (unstabilized) methods.
}
