% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateIPTW.R
\name{unstabilizedIPTW}
\alias{unstabilizedIPTW}
\title{Function to Compute Horwitz-Thompson}
\usage{
unstabilizedIPTW(outcome, treatment, prob_treated, ...)
}
\arguments{
\item{outcome}{vector of outcome values}

\item{treatment}{vector of treatment values}

\item{prob_treated}{vectof of propensity scores}

\item{...}{dots}
}
\description{
Computes estimates for each individual by IPTW (unstabilized) methods.
}
