% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDR.R
\name{estimateDRIPTW}
\alias{estimateDRIPTW}
\title{Estimate ATE with Doubly Robust IPTW.}
\usage{
estimateDRIPTW(data, outcome_formula, outcome_model_method = "logistic",
  treatment_formula, treatment_model_method = "logistic", ...)
}
\arguments{
\item{data}{The dataframe. This will be coerced from "tbl_df" to data.frame.}

\item{outcome_formula}{A two-part formula for the outcome regression.}

\item{outcome_model_method}{currently only supported "logistic" for
logit-link binomial GLM.}

\item{treatment_formula}{A two-part formula for the treatment model.}

\item{treatment_model_method}{currently only supported "logistic" for
logit-link binomial GLM.}

\item{...}{additional arguments, especially those passed to
\code{\link[geex]{m_estimate}} or the modeling function}
}
\description{
Fits a parametric model and estimates ATE via IPTW with Wald-type confidence
intervals from the empirical sandwich standard error estimates.
}
