% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeFunGF.R
\name{estimateGF}
\alias{estimateGF}
\title{Estimate ATE with IPTW}
\usage{
estimateGF(data, formula, model_method = "logistic", treatment_var_name)
}
\arguments{
\item{data}{the dataframe. Will be coerced from "tbl_df" to data.frame.}

\item{formula}{Three-part formula: Outcome | Treatment ~ model_predictors. Will be coerced to object of type Formula.}

\item{model_method}{currently only supported "logistic" for logit-link binomial GLM.}

\item{treatment_var_name}{the name of the treatment variable. Must line up with formula.}
}
\description{
Fits a parametric model and estimates ATE via IPTW with Wald-type confidence
intervals from the empirical sandwich standard error estimates.
}
