% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeFunIPW.R
\name{estimateIPTW}
\alias{estimateIPTW}
\title{Estimate ATE with IPTW}
\usage{
estimateIPTW(data, formula, model_method = "logistic",
  weight_type = "unstabilized", ...)
}
\arguments{
\item{data}{the dataframe. Will be coerced from "tbl_df" to data.frame.}

\item{formula}{Three-part formula: Outcome | Treatment ~ model_predictors. Will be coerced to object of type Formula.}

\item{model_method}{currently only supported "logistic" for logit-link binomial GLM.}

\item{weight_type}{Currently only supports "unstabilized"}

\item{...}{additional args}
}
\description{
Fits a parametric model and estimates ATE via IPTW with Wald-type confidence
intervals from the empirical sandwich standard error estimates.
}
